As the trunk has been modified a lot in the meantime there are too many differences to continue on the old branch. As I don't know any way to update the branch to match the trunk, I create a new branch. Now we will at least have the current trunk state...
Only one problem is left:
there are several branches 'ReadyForRelease', but not yet merged. As long as this branch OR the other branches aren't merged, these are not yet translated...

Michael, when you are back, could you please look at the 'ReadyForRelease' branches? When we merged everything into trunk immediatly this had to happen ASAP, as no other releases could be made until the change was 'correct'. Now some branches are waiting a long time without being released, causing the kind of issues I mentioned above...