Update issue 4893 Status: ReadyForRelease Labels: Priority-High Summary: Hide the placeholders of ads, instead of removing them
I thought it might be a good idea to make this optional. The web has many many web pages, and there probably will be some that use element.style = "X", which overrules our purging method.
However, I'm not capable of searching all webpages, so I guess it is best to make it a non-default option, and await the (few or many) reports. Then we'll know better what to do than by guessing it.