Update issue 5914 Status: NeedsReview
Add the shell of the webRequest blocking code to background.html.
This shell registers all the handlers that the real code would, but the
handlers either don't do anything, or explicitly throw an error message
so we can verify that AdBlock falls back to 'old' style correctly.

Things you should be able to do at this point:
1. Start AdBlock, have it fall back to 'old' style, and verify that
blocking works properly (I tend to search for 'boots' on thepiretabay.org
and verify no banner ads on the sides of the page)
2. Add 'experimental' permission to manifest.json, start AdBlock, and
verify that AdBlock still falls back to the 'old' style and blocks
properly.  background.html's Console should show that it fell back because
of the explicitly thrown error in onBeforeRequestHandler.
3. comment out the throw in onBeforeRequestHandler, add 'experimental'
permission, and verify that AdBlock stays in 'new' style: hiding works on
e.g. slashdot.org, but blocking on thepiratebay fails.

Next branch will implement the blocking itself.