Update issue 5915 First step: tab and frame URL tracking.
When a tab is created, onBeforeRequest tells you the url being fetched.
Similarly when a subframe is created, onBeforeRequest tells you the subframe's URL.
But when an image/object/video/etc is fetched, you aren't told the URL of the tab
or frame that made the request.

Since AdBlock needs to know 3 URLS (page URL, frame URL for third party detection,
resource URL), we must track the page and frame URLs ourselves.

When a tab/frame is created, we record the URL.
When a frame navigates, we update its info.
When a tab navigates, we wipe out all frame info and record the new tab info.
When a tab closes, we wipe out all its frame info.

Verified that with 'experimental' permission, this code correctly fills frameData
object with URLs, and handles navigation and tab closing correctly.