Update issue 6248 Throwing in handlerBehaviorChanged() while I'm at it.
You can pass function() { alert('hi'); } to handlerBehaviorChanged
to make it alert every time it's called.  It should run once upon
startup, once upon unpause, once upon filters change.

Upon install it runs three or four times, which is less than the quota
of 10-in-a-short-time that I think is built into Chrome.

(handlerBehaviorChanged needs to be called every time we need the in-memory
cache cleared, e.g. we paused, an ad was loaded, we unpaused, now the cache
has the ad even though we want to block it.)