Update issue 6909 Status: NeedsReview Really, there is some connection between "train" and "resourceblock" as once again I was in a train and once again resourceblock is the component that gets an update (probably because I don't need an internet connection for testing resourceblock).
This time: make search faster. We were totally wrong. Although I applied the suggestion of joining the keywords into a string (just thought about there is one character that can't be used anywhere, logically: \t AKA tab). However, this turned out to be a no-op. I left it in because it doesn't harm, but well, it was still terribly slow.
However, one thing that proved to be the slowdown wasn't the matching or such. Not at all: it was $(el).hide().
el.style.setProperty("display", "none", "") was just a little bit faster. However, setting a class instead is very fast. Now, with 1000 test resources, it takes < 1 second with multiple wildcards and keywords. Thus incremental search is on at all times now.