Update issue 6988 Status: NeedsReview First attempt :P I personally think this message is "soft" enough not to instantly cause a heart attack, but still tells them what to do.
It's probably going to become a FAQ item ("I see ads on every site I visit, what now?"), so the message probably will change. However, it should still mention something about the reasons, as we can't just say "Hey, we think you can better read [this] page" without a reason.
"Not all pages on the internet contain advertisements. The fact that you see them may indicate that your computer contains some unwanted software. Please have a look at [this] page to see how to get rid of it."
(speaking of which: is there a word other than "item" or "thing" that contains both "extension" and "executables"? I used software here, but extensions aren't the kind of software as it's mostly referred to.)

I also updated the step# IDs in the code, so we don't have to change them all the time when we add/remove a step.